predict DamageOnLand
