+++ Content.Trauma.Common/Mail/MailFragileEvent.cs
@@ -1,0 +1,8 @@
+namespace Content.Trauma.Common.Mail;
+
+/// <summary>
+/// Raised on mail to check if it is fragile.
+/// TODO MAIL: remove with nyano mail removal
+/// </summary>
+[ByRefEvent]
+public record struct MailFragileEvent(bool Fragile = false);
